fixed some bugs in Game.js, App.js
