+++ client/src/App.js
@@ -8,12 +8,10 @@ import RegisterUser from "./Components/login/RegisterUser";
 import Users from "./Components/users/Users";
 import LoginUser from "./Components/login/LoginUser";
 import LeaderBoard from "./Components/Game/LeaderBoard";
-import currencyFormatter from "currency-formatter";
-import swal from "sweetalert2";
 import axios from "axios";
 import "./Views/App.css";
 
-//very fragile
+//very fragile - if Material UI changes the classNames, the app breaks.
 import JssProvider from 'react-jss/lib/JssProvider';
 import { createGenerateClassName } from '@material-ui/core/styles';
 
@@ -24,64 +22,6 @@ const generateClassName = createGenerateClassName({
   dangerouslyUseGlobalCSS: true,
   productionPrefix: 'c',
 });
-
-const TMDB_KEY = process.env.REACT_APP_TMDB_API_KEY;
-const baseURL = `http://image.tmdb.org/t/p/w185`;
-const idArr = [
-  "tt0111161",
-  "tt0068646",
-  "tt0071562",
-  "tt0468569",
-  "tt0050083",
-  "tt0108052",
-  "tt0110912",
-  "tt0167260",
-  "tt0060196",
-  "tt0137523",
-  "tt0120737",
-  "tt1615160",
-  "tt0451279",
-  "tt4574334",
-  "tt5580390",
-  "tt3501632",
-  "tt0109830",
-  "tt0080684",
-  "tt1375666",
-  "tt0167261",
-  "tt0073486",
-  "tt0099685",
-  "tt0133093"
-];
-const idArr1 = [
-  "tt0047478",
-  "tt0076759",
-  "tt0120382",
-  "tt0107290",
-  "tt0477348",
-  "tt0395169",
-  "tt1201607",
-  "tt0264464",
-  "tt1856101",
-  "tt0435761",
-  "tt1675434",
-  "tt0120815",
-  "tt0361748",
-  "tt0180093",
-  "tt0054215",
-  "tt1485796",
-  "tt2283362",
-  "tt1259528",
-  "tt2527336",
-  "tt2380307",
-  "tt4765284",
-  "tt5052448",
-  "tt1396484",
-  "tt5463162",
-  "tt4154756",
-  "tt3778644",
-  "tt1677720"
-];
-
 
 class App extends React.Component {
   constructor(props) {
@@ -130,7 +70,6 @@ class App extends React.Component {
     axios
       .get("/users/userinfo")
       .then(res => {
-        // Current
         let user = res.data.userInfo;
         if (!this.state.loggedIn) {
           this.setState({
@@ -156,19 +95,6 @@ class App extends React.Component {
       })
       .catch(err => {
         console.error("Error Getting Blogs:", err);
-      });
-  };
-
-  getUserScore = () => {
-    axios
-      .get("/users/getcurrentscore")
-      .then(res => {
-        this.setState({
-          score: res.data.data.points
-        });
-      })
-      .catch(err => {
-        console.error("Error getting user score:", err);
       });
   };
 
@@ -235,165 +161,6 @@ class App extends React.Component {
       });
   };
 
-  postScore = () => {
-    const { user } = this.state;
-    const { score } = this.state;
-
-    axios
-      .patch("users/score_update", {
-        points: score,
-        id: user.id
-      })
-      .then(res => {
-
-      })
-      .catch(error => {
-        console.error(error);
-      });
-  };
-
-  getTwoMovies = () => {
-    this.setState({
-      hasBeenClicked: false
-    });
-
-    let randomMovieID1 = `${idArr[Math.floor(Math.random() * idArr.length)]}`;
-
-    let randomMovieID2 = `${idArr1[Math.floor(Math.random() * idArr1.length)]}`;
-    axios
-      .get(
-        `http://api.themoviedb.org/3/movie/${randomMovieID1}?api_key=${TMDB_KEY}`
-      )
-      .then(response => {
-        this.setState({
-          movie1: response,
-          movie1Revenue: response.data.revenue,
-          movie1Budget: response.data.budget,
-          // eslint-disable-next-line
-          movie1MoneyEarned: eval(response.data.revenue - response.data.budget)
-        });
-      })
-      .then(
-        axios
-          .get(
-            `http://api.themoviedb.org/3/movie/${randomMovieID2}?api_key=${TMDB_KEY}`
-          )
-          .then(response => {
-            this.setState({
-              movie2: response,
-              movie2Revenue: response.data.revenue,
-              movie2Budget: response.data.budget,
-              // eslint-disable-next-line
-              movie2MoneyEarned: eval(
-                response.data.revenue - response.data.budget
-              )
-            });
-          })
-          .then(() => {
-            this.setState({
-              winner:
-                this.state.movie1MoneyEarned >= this.state.movie2MoneyEarned
-                  ? this.state.movie1.data
-                  : this.state.movie2.data,
-              loser:
-                this.state.movie1MoneyEarned <= this.state.movie2MoneyEarned
-                  ? this.state.movie1.data
-                  : this.state.movie2.data
-            });
-          })
-          .catch(error => {
-            console.error(error);
-          })
-      )
-      .catch(error => {
-        console.error(error);
-      });
-  };
-
-  getWinner = e => {
-    if (!this.state.winner && !this.state.loser) {
-      window.location.reload();
-    }
-    e.preventDefault();
-    const {
-      winner,
-      loser,
-      user,
-      movie1MoneyEarned,
-      movie2MoneyEarned,
-      hasBeenClicked
-    } = this.state;
-
-    let diff = movie1MoneyEarned - movie2MoneyEarned;
-    let currentTargetJoined = e.currentTarget.innerText.split(" ").join("").trim();
-    let winnerJoined = winner.original_title.split(" ").join("").trim();
-    let loserJoined = loser.original_title.split(" ").join("").trim();
-    debugger;
-    if (currentTargetJoined === winnerJoined && !hasBeenClicked) {
-      this.setState({
-        // eslint-disable-next-line
-        score: (this.state.score += 10),
-        hasBeenClicked: true
-      });
-      if (hasBeenClicked) {
-        this.setState({
-          // eslint-disable-next-line
-          score: (this.state.score += 0)
-        });
-      }
-      swal({
-        customClass: "animated rubberBand",
-        html: `<div id="win-win"><span id="swal-message-right"><h6 className="swal-alert">Yes, ${
-          user.firstname
-        }! ${winner.original_title} made approximately <h2> ${currencyFormatter.format(Math.abs(diff), {
-          code: "USD"
-        })}</h2> more than ${
-          loser.original_title
-        }</h6 className="swal-alert"></span></div>`,
-        background: `#eee url(${baseURL}${winner.backdrop_path}) space`,
-        backdrop: `
-        rgba(0,255,0,0.5)`,
-        imageWidth: 400,
-        imageHeight: 200,
-        imageAlt: "Custom image",
-        animation: false,
-        confirmButtonColor: "#02182b",
-        showLoaderOnConfirm: true
-      }).then(this.getTwoMovies())
-    }
-    if (currentTargetJoined === loserJoined) {
-      this.setState({
-        hasBeenClicked: true
-      });
-
-
-      swal({
-        customClass: "animated headShake",
-        html: `<div id="win-win"><span id="swal-message"><h6>Sorry, ${
-          user.firstname
-        }... ${
-          e.target.title
-        }  made approximately <h2>${currencyFormatter.format(Math.abs(diff), {
-          code: "USD"
-        })}</h2> less than ${
-          winner.original_title
-        }.</h6></span></div>`,
-        // background: `#eee url(${baseURL}${loser.backdrop_path}) space`,
-        backdrop: `
-        rgba(255,0,0,0.5)`,
-        imageWidth: 400,
-        imageHeight: 200,
-        imageAlt: "Custom image",
-        animation: false,
-        confirmButtonColor: "#02182b",
-        showLoaderOnConfirm: true
-      }).then(this.getTwoMovies())
-    }
-    if (!hasBeenClicked) {
-      this.postScore();
-    }
-  };
-
   componentDidMount() {
     axios
       .get("/users/userinfo")
@@ -406,18 +173,6 @@ class App extends React.Component {
       .catch(err => {
         console.error("Error Getting User Info:", err);
       });
-
-    axios
-      .get("/users/getcurrentscore")
-      .then(res => {
-        this.setState({
-          score: res.data.data.points
-        });
-      })
-      .then(this.getLeaderBoard())
-      .catch(err => {
-        console.error("Error Getting Score:", err);
-      });
   }
 
   render() {
@@ -427,18 +182,12 @@ class App extends React.Component {
       username,
       password,
       message,
-      score,
       leaderBoardData,
-      allBlogs,
-      movie1,
-      movie2,
-      winner,
-      loser,
-      movie1MoneyEarned,
-      movie2MoneyEarned,
-      hasBeenClicked
+      allBlogs
     } = this.state;
+
     const { classes } = this.props;
+
     const {
       handleInputChange,
       submitLoginForm,
@@ -449,9 +198,7 @@ class App extends React.Component {
       logOut,
       handleClick,
       getLeaderBoard,
-      getAllBlogPosts,
-      getTwoMovies,
-      getWinner
+      getAllBlogPosts
     } = this;
 
     return (
@@ -464,7 +211,6 @@ class App extends React.Component {
           getUserInfo={getUserInfo}
           logOut={logOut}
           classes={classes}
-          score={score}
           getUserScore={getUserScore}
         />
         <Switch>
@@ -486,7 +232,6 @@ class App extends React.Component {
                 password={password}
                 message={message}
                 loggedIn={loggedIn}
-                score={score}
               />
             )}
           />
@@ -506,7 +251,6 @@ class App extends React.Component {
             render={props => (
               <Users
                 {...props}
-                score={score}
                 getUserInfo={getUserInfo}
                 currentUser={user}
                 getUserScore={getUserScore}
@@ -522,21 +266,8 @@ class App extends React.Component {
             render={props => (
               <Game
                 {...props}
-                originalScore={score}
                 loggedIn={loggedIn}
                 currentUser={user}
-                getUserScore={getUserScore}
-                getLeaderBoard={getLeaderBoard}
-                movie1={movie1}
-                movie2={movie2}
-                winner={winner}
-                loser={loser}
-                movie1MoneyEarned={movie1MoneyEarned}
-                movie2MoneyEarned={movie2MoneyEarned}
-                score={score}
-                hasBeenClicked={hasBeenClicked}
-                getTwoMovies={getTwoMovies}
-                getWinner={getWinner}
               />
             )}
           />
@@ -545,7 +276,6 @@ class App extends React.Component {
             render={props => (
               <Favorites
                 {...props}
-                score={score}
                 loggedIn={loggedIn}
                 currentUser={user}
               />
@@ -556,7 +286,6 @@ class App extends React.Component {
             render={props => (
               <LeaderBoard
                 {...props}
-                score={score}
                 loggedIn={loggedIn}
                 currentUser={user}
                 data={leaderBoardData}
